Issue #8: Use shellTextArea from shell-emulator Module
